+++ template/src/components/participants/Participant.tsx
@@ -19,7 +19,7 @@ import {isMobileUA, isWebInternal} from '../../utils/common';
 import ActionMenu, {ActionMenuItem} from '../../atoms/ActionMenu';
 import Spacer from '../../atoms/Spacer';
 import useRemoteEndCall from '../../utils/useRemoteEndCall';
-import {useChatMessages} from '../chat-messages/useChatMessages';
+import {useChatMessages} from '../chat-messages/useChatMessages0';
 import LocalVideoMute from '../../subComponents/LocalVideoMute';
 import LocalAudioMute from '../../subComponents/LocalAudioMute';
 import RemoveMeetingPopup from '../../subComponents/RemoveMeetingPopup';
