+++ template/src/components/participants/UserActionMenuOptions.tsx
@@ -17,7 +17,7 @@ import useRemoteRequest, {
   REQUEST_REMOTE_TYPE,
 } from '../../utils/useRemoteRequest';
 import ActionMenu, {ActionMenuItem} from '../../atoms/ActionMenu';
-import {useChatMessages} from '../chat-messages/useChatMessages';
+import {useChatMessages} from '../chat-messages/useChatMessages0';
 import {useLiveStreamDataContext} from '../contexts/LiveStreamDataContext';
 import useRemoteEndCall from '../../utils/useRemoteEndCall';
 import LiveStreamContext from '../livestream/LiveStreamContext';
